@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: "2.9.2"
+  sdk: "2.13.1"
 
 dependencies:
   flutter:
@@ -28,6 +28,9 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
+  lottie: ^1.0.1
+  google_fonts: ^2.1.0
+  
 
 dev_dependencies:
   flutter_test:
